Configure semantic-release
This adds and sets up the `semantic-release` package to use commits and travis to do automated semantic versioning and publishing to npm for scratch-parser.

It also adds a readme section about committing, and using `git cz` to handle formatting of commits in a way that is parseable and deployable by `semantic-release`.

This setup is based on the existing release setup in `eslint-config-scratch`.