refactor($lib, $index, $schema): Refactor to use promises to accomodate the use of the JSZip library
This library has been majorly re-factored to use promises to accomodate the use of the JSZip library to unpack a provided buffer. Callback style usage of this library is still supported, with the only change to usage being that any piece of code that must relies on the functions provided by this library to completely finish executing must be called from within the function callback. An example of this is using test.end() to ensure that a test and all its subprocesses have finished executing.

BREAKING CHANGE: Anything that needs to wait for any of these functions to finish before
proceeding must be called in function callback. Calling test.end() on a test is a good example of this.